eureka-backend/requirement.txt
fastapi
uvicorn
pymongo
python-dotenv
bcrypt
openai
torch
sentence-transformers
pandas
openpyxl
pandasai
bcrypt
azure-storage-blob
pinecone